Break out pricing logic into a separate class
